+++ internal/controller/operator/factory/vmagent/podscrape.go
@@ -57,13 +57,21 @@ func generatePodScrapeConfig(
 	relabelings = addSelectorToRelabelingFor(relabelings, "pod", m.Spec.Selector)
 
 	// Filter targets based on correct port for the endpoint.
-	if ep.Port != "" {
+	switch {
+	case ep.Port != nil && len(*ep.Port) > 0:
 		relabelings = append(relabelings, yaml.MapSlice{
 			{Key: "action", Value: "keep"},
 			{Key: "source_labels", Value: []string{"__meta_kubernetes_pod_container_port_name"}},
-			{Key: "regex", Value: ep.Port},
+			{Key: "regex", Value: *ep.Port},
 		})
-	} else if ep.TargetPort != nil {
+
+	case ep.PortNumber != nil && *ep.PortNumber != 0:
+		relabelings = append(relabelings, yaml.MapSlice{
+			{Key: "action", Value: "keep"},
+			{Key: "source_labels", Value: []string{"__meta_kubernetes_pod_container_port_number"}},
+			{Key: "regex", Value: *ep.PortNumber},
+		})
+	case ep.TargetPort != nil:
 		if ep.TargetPort.StrVal != "" {
 			relabelings = append(relabelings, yaml.MapSlice{
 				{Key: "action", Value: "keep"},
@@ -124,10 +132,10 @@ func generatePodScrapeConfig(
 		})
 	}
 
-	if ep.Port != "" {
+	if ep.Port != nil && len(*ep.Port) > 0 {
 		relabelings = append(relabelings, yaml.MapSlice{
 			{Key: "target_label", Value: "endpoint"},
-			{Key: "replacement", Value: ep.Port},
+			{Key: "replacement", Value: *ep.Port},
 		})
 	} else if ep.TargetPort != nil && ep.TargetPort.String() != "" {
 		relabelings = append(relabelings, yaml.MapSlice{