+++ test/e2e/watchnamespace/prometheus_converter_test.go
@@ -8,6 +8,7 @@ import (
 	. "github.com/onsi/gomega"
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -35,7 +36,7 @@ var _ = Describe("VM Operator", func() {
 				},
 				Spec: monitoringv1.PodMonitorSpec{
 					PodMetricsEndpoints: []monitoringv1.PodMetricsEndpoint{{
-						Port: "9999",
+						Port: ptr.To("9999"),
 					}},
 				},
 			},